+++ actionpack/test/dispatch/system_testing/driver_test.rb
@@ -37,6 +37,24 @@ class DriverTest < ActiveSupport::TestCase
     assert_equal ({ url: "http://example.com/wd/hub" }), driver.instance_variable_get(:@options)
   end
 
+  test "initializing the driver with a poltergeist" do
+    driver = assert_deprecated do
+      ActionDispatch::SystemTesting::Driver.new(:poltergeist, screen_size: [1400, 1400], options: { js_errors: false })
+    end
+    assert_equal :poltergeist, driver.instance_variable_get(:@name)
+    assert_equal [1400, 1400], driver.instance_variable_get(:@screen_size)
+    assert_equal ({ js_errors: false }), driver.instance_variable_get(:@options)
+  end
+
+  test "initializing the driver with a webkit" do
+    driver = assert_deprecated do
+      ActionDispatch::SystemTesting::Driver.new(:webkit, screen_size: [1400, 1400], options: { skip_image_loading: true })
+    end
+    assert_equal :webkit, driver.instance_variable_get(:@name)
+    assert_equal [1400, 1400], driver.instance_variable_get(:@screen_size)
+    assert_equal ({ skip_image_loading: true }), driver.instance_variable_get(:@options)
+  end
+
   test "initializing the driver with a cuprite" do
     driver = ActionDispatch::SystemTesting::Driver.new(:cuprite, screen_size: [1400, 1400], options: { js_errors: false })
     assert_equal :cuprite, driver.instance_variable_get(:@name)